+++ frontend/jest.config.ts
@@ -10,6 +10,7 @@ const config: Config = {
     '^.+\\.svg$': '<rootDir>/src/__mocks__/svg.ts',
     '^@/(.*)$': '<rootDir>/src/$1',
     '^import.meta.env$': '<rootDir>/src/env.ts',
+    '\\.(css)$': '<rootDir>/src/__mocks__/styleMock.ts',
   },
   watchPathIgnorePatterns: ['node_modules'],
   setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],