+++ frontend/src/features/map_planning/routes/MapWrapper.tsx
@@ -9,14 +9,14 @@ import useMapStore from '../store/MapStore';
 import { handleRemoteAction } from '../store/RemoteActions';
 import { mapEditorSteps, tourOptions } from '../utils/EditorTour';
 import { ReadOnlyModeContextProvider } from '../utils/ReadOnlyModeContext';
-import { LayerType, LayerDto, GuidedToursDto } from '@/api_types/definitions';
+import { LayerType, LayerDto } from '@/api_types/definitions';
 import { createAPI } from '@/config/axios';
 import { QUERY_KEYS } from '@/config/react_query';
 import { PolygonGeometry } from '@/features/map_planning/types/PolygonTypes';
 import { errorToastGrouped } from '@/features/toasts/groupedToast';
 import { useSafeAuth } from '@/hooks/useSafeAuth';
 import { useQuery } from '@tanstack/react-query';
-import { useRef, useEffect, useState } from 'react';
+import { useRef, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ShepherdOptionsWithType, ShepherdTour } from 'react-shepherd';
 
@@ -228,15 +228,14 @@ export function MapWrapper() {
   const mapData = useInitializeMap();
   useMapUpdates();
 
-  const [status, setStatus] = useState<GuidedToursDto>();
-  useTourStatus(setStatus);
+  const tourStatus = useTourStatus();
 
   if (!mapData) {
     return null;
   }
 
   let steps: ShepherdOptionsWithType[] = [];
-  if (status && !status.editor_tour_completed) {
+  if (tourStatus && !tourStatus.editor_tour_completed) {
     steps = mapEditorSteps;
   }