+++ frontend/src/features/seeds/api/findPlantById.ts
@@ -1,8 +1,15 @@
 import { PlantsSummaryDto } from '@/api_types/definitions';
 import { createAPI } from '@/config/axios';
 
-export const findPlantById = async (id: number): Promise<PlantsSummaryDto> => {
+export const findPlantById = async (id: number | undefined): Promise<PlantsSummaryDto> => {
   const http = createAPI();
+
+  if (id === undefined) {
+    return new Promise((resolve, reject) => {
+      reject();
+    });
+  }
+
   try {
     const response = await http.get<PlantsSummaryDto>(`/api/plants/${id}`);
     return response.data;